blog page added

+++ src/Pages/Home/MyToys/MyToys.jsx
@@ -11,24 +11,37 @@ const MyToys = () => {
     const [toys, setToys] = useState(loadedToys);
 
     const handleDeleteToy = (_id) => {
-        fetch(`http://localhost:5000/posts/${_id}`, {
-            method: 'DELETE'
-        })
-        .then(res => res.json())
-        .then(data => {
-            console.log(data);
-            if (data.deletedCount > 0) {
-                Swal.fire({
-                    position: ' center',
-                    icon: 'success',
-                    title: 'Deleted Succesfull',
-                    showConfirmButton: false,
-                    timer: 1500
-                  })
-                const remaining = toys.filter(toy => toy._id !== _id);
-                setToys(remaining)
+        Swal.fire({
+            title: 'Are you sure?',
+            text: "You won't be able to revert this!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, delete it!'
+        }).then((result) => {
+            if (result.isConfirmed) {
+                fetch(`http://localhost:5000/posts/${_id}`, {
+                    method: 'DELETE'
+                })
+                    .then(res => res.json())
+                    .then(data => {
+                        console.log(data);
+                        if (data.deletedCount > 0) {
+                            Swal.fire({
+                                position: ' center',
+                                icon: 'success',
+                                title: 'Deleted Succesfull',
+                                showConfirmButton: false,
+                                timer: 1500
+                            })
+                            const remaining = toys?.filter(toy => toy._id !== _id);
+                            setToys(remaining)
+                        }
+                    })
             }
         })
+
     }
 
     return (
@@ -47,7 +60,7 @@ const MyToys = () => {
                 </div>
                 <div>
                     {
-                        toys.map(toy => <MyToy handleDeleteToy={handleDeleteToy} key={toy._id} toy={toy}/>)
+                        toys?.map(toy => <MyToy handleDeleteToy={handleDeleteToy} key={toy._id} toy={toy} />)
                     }
                 </div>
             </div>
